+++ src/main/java/com/javacourse/project/hibernateAndJpa/Business/ICityService.java
@@ -10,5 +10,5 @@ public interface ICityService {
 	void add(City city);
 	void update(City city);
 	void delete(City city);
-	City getById(int id);
+	City getById(Long id);
 }
