+++ src/main/java/com/javacourse/project/hibernateAndJpa/Dal/ICityDal.java
@@ -7,5 +7,5 @@ public interface ICityDal {
 	void add(City city);
 	void update(City city);
 	void delete(City city);
-	City getById(int id);
+	City getById(Long id);
 }
